+++ packages/base/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - new attribute `model.realPath` if added to for each model collection item returned by `argv.getServerConfiguration()`.
+ Contains a full file system path to the model folder
 
 ### Changed
  - fix ESLint warnings (mostly about let -> const)